requirements.txt
pycryptodome~=3.23

pwntools~=4.14
